3lab/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-jose[cryptography]  # Для JWT
passlib                    # Хеширование паролей
pydantic[email]
bcrypt
jinja2
Pillow
numpy
opencv-python-headless
python-multipart
celery
redis
